pyproject.toml
[project]
name = "monitor-copilot-enterprise"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
